Rework KMMetadata to Metadata and internalize conversion functions.
