+++ kmp-grpc-core/src/androidJvmCommon/kotlin/io/github/timortel/kmpgrpc/core/message/DataType.kt
