--- kmp-grpc-core/src/androidJvmCommon/kotlin/io/github/timortel/kmpgrpc/core/statusutil.kt
@@ -1,6 +1,0 @@
-package io.github.timortel.kmpgrpc.core
-
-import io.grpc.Status
-
-val KMStatus.jvmStatus: Status
-    get() = Status.fromCodeValue(code.value).withDescription(statusMessage)
