+++ kmp-grpc-core/src/iosJvmCommon/kotlin/io/github/timortel/kmpgrpc/core/io/writer.kt
@@ -2,7 +2,7 @@ package io.github.timortel.kmpgrpc.core.io
 
 import io.github.timortel.kmpgrpc.core.message.KMMessage
 
-fun writeKMMessage(
+internal fun writeKMMessage(
     stream: CodedOutputStream,
     fieldNumber: Int,
     msg: KMMessage,
@@ -15,7 +15,7 @@ fun writeKMMessage(
     writeMessage(msg, stream)
 }
 
-fun writeMessageList(
+internal fun writeMessageList(
     stream: CodedOutputStream,
     fieldNumber: Int,
     values: List<KMMessage>,
@@ -25,7 +25,7 @@ fun writeMessageList(
     values.forEach { writeKMMessage(stream, fieldNumber, it, requiredSize(it), writeMessage) }
 }
 
-fun <T> writeArray(
+internal fun <T> writeArray(
     stream: CodedOutputStream,
     fieldNumber: Int,
     values: Collection<T>,
@@ -44,5 +44,28 @@ fun <T> writeArray(
         values.forEach(writeNoTag)
     } else {
         values.forEach { writeTag(fieldNumber, it) }
+    }
+}
+
+fun <K, V> writeMap(
+    stream: CodedOutputStream,
+    fieldNumber: Int,
+    map: Map<K, V>,
+    getKeySize: (fieldNumber: Int, key: K) -> Int,
+    getValueSize: (fieldNumber: Int, value: V) -> Int,
+    writeKey: CodedOutputStream.(fieldNumber: Int, K) -> Unit,
+    writeValue: CodedOutputStream.(fieldNumber: Int, V) -> Unit
+) {
+    val tag = wireFormatMakeTag(fieldNumber, WireFormat.LENGTH_DELIMITED)
+    map.forEach { (key, value) ->
+        //Write tag
+        stream.writeInt32NoTag(tag)
+        //Write the size of the message
+        val msgSize =
+            getKeySize(kMapKeyFieldNumber, key) + getValueSize(kMapValueFieldNumber, value)
+        stream.writeInt32NoTag(msgSize)
+        //Write fields
+        writeKey(stream, kMapKeyFieldNumber, key)
+        writeValue(stream, kMapValueFieldNumber, value)
     }
 }
