+++ kmp-grpc-core/src/jsMain/kotlin/io/github/timortel/kmpgrpc/core/internal/UnaryResponseImpl.kt
@@ -1,11 +1,12 @@
 package io.github.timortel.kmpgrpc.core.internal
 
-import io.github.timortel.kmpgrpc.core.Metadata
-import io.github.timortel.kmpgrpc.core.MethodDescriptor
-import io.github.timortel.kmpgrpc.core.Status
-import io.github.timortel.kmpgrpc.core.UnaryResponse
+import io.github.timortel.kmpgrpc.core.JsMetadata
+import io.github.timortel.kmpgrpc.core.external.Metadata
+import io.github.timortel.kmpgrpc.core.external.MethodDescriptor
+import io.github.timortel.kmpgrpc.core.external.Status
+import io.github.timortel.kmpgrpc.core.external.UnaryResponse
 
-class UnaryResponseImpl<RESP>(
+internal class UnaryResponseImpl<RESP>(
     val resp: RESP,
     val metadata: Metadata,
     val status: Status,
@@ -13,7 +14,7 @@ class UnaryResponseImpl<RESP>(
 ) : UnaryResponse<RESP> {
     override fun getResponseMessage(): RESP = resp
 
-    override fun getMetadata(): Metadata = metadata
+    override fun getMetadata(): JsMetadata = metadata
 
     override fun getStatus(): Status = status
