+++ kmp-grpc-core/src/jsMain/kotlin/io/github/timortel/kmpgrpc/core/message/MessageDeserializer.kt
@@ -1,6 +1,6 @@
 package io.github.timortel.kmpgrpc.core.message
 
-import io.github.timortel.kmpgrpc.core.util.native
+import io.github.timortel.kmpgrpc.core.native
 import org.khronos.webgl.Uint8Array
 
 actual interface MessageDeserializer<T : KMMessage> {
