--- kmp-grpc-core/src/jsMain/kotlin/io/github/timortel/kmpgrpc/core/metadata_util.kt
@@ -1,8 +1,0 @@
-package io.github.timortel.kmpgrpc.core
-
-import kotlin.js.json
-
-val KMMetadata.jsMetadata: Metadata
-    get() = json(
-        *metadataMap.entries.map { (key, value) -> key to value }.toTypedArray()
-    ).unsafeCast<Metadata>()
