+++ kmp-grpc-core/src/jsMain/kotlin/io/github/timortel/kmpgrpc/core/stub/JsStub.kt
@@ -1,20 +1,20 @@
 package io.github.timortel.kmpgrpc.core.stub
 
 import io.github.timortel.kmpgrpc.core.KMChannel
-import io.github.timortel.kmpgrpc.core.KMMetadata
+import io.github.timortel.kmpgrpc.core.Metadata
 import io.github.timortel.kmpgrpc.core.util.TimeUnit
 
 interface JsStub<S : KMStub<S>> {
 
     val channel: KMChannel
-    val callOptions: KMMetadata
+    val callOptions: Metadata
 
-    fun build(channel: KMChannel, callOptions: KMMetadata): S
+    fun build(channel: KMChannel, callOptions: Metadata): S
 
     fun withDeadlineAfter(duration: Long, unit: TimeUnit): S {
-        val newMetadata = callOptions.copy()
-        newMetadata["deadline"] = (duration * unit.toMilliFactor).toString()
-
-        return build(channel, newMetadata)
+        return build(
+            channel = channel,
+            callOptions = callOptions.withEntry("deadline", (duration * unit.toMilliFactor).toString())
+        )
     }
 }
