+++ kmp-grpc-plugin/src/main/java/io/github/timortel/kmpgrpc/plugin/sourcegeneration/constants/lib_packages.kt
@@ -7,3 +7,4 @@ const val PACKAGE_MESSAGE = "$PACKAGE_BASE.message"
 const val PACKAGE_STUB = "$PACKAGE_BASE.stub"
 const val PACKAGE_RPC = "$PACKAGE_BASE.rpc"
 const val PACKAGE_UTIL = "$PACKAGE_BASE.util"
+const val PACKAGE_EXTERNAL = "$PACKAGE_BASE.external"
