+++ kmp-grpc-plugin/src/main/java/io/github/timortel/kmpgrpc/plugin/sourcegeneration/constants/library_fields.kt
@@ -3,13 +3,11 @@ package io.github.timortel.kmpgrpc.plugin.sourcegeneration.constants
 import com.squareup.kotlinpoet.ClassName
 import com.squareup.kotlinpoet.MemberName
 
-val kmMetadata = ClassName(PACKAGE_BASE, "KMMetadata")
+val kmMetadata = ClassName(PACKAGE_BASE, "Metadata")
 val kmStub = ClassName(PACKAGE_STUB, "KMStub")
 val kmChannel = ClassName(PACKAGE_BASE, "KMChannel")
-val kmAndroidJVMStub = ClassName(PACKAGE_STUB, "AndroidJvmKMStub")
 val kmTimeUnit = ClassName(PACKAGE_UTIL, "TimeUnit")
 
-val JSImpl = ClassName(PACKAGE_MESSAGE, "JSImpl")
 val MessageDeserializer = ClassName(PACKAGE_MESSAGE, "MessageDeserializer")
 val MessageCompanion = ClassName(PACKAGE_MESSAGE, "KMMessageCompanion")
 
@@ -18,7 +16,6 @@ val kmEnum = ClassName(PACKAGE_MESSAGE, "KmEnum")
 
 val unknownField = ClassName(PACKAGE_MESSAGE, "UnknownField")
 
-val writeKMMessage = MemberName(PACKAGE_IO, "writeKMMessage")
 val readKMMessage = MemberName(PACKAGE_IO, "readKMMessage")
 
 val readUnknownField = MemberName(PACKAGE_IO, "readUnknownField")
@@ -39,6 +36,5 @@ val iosServerSideStreamingCallImplementation = MemberName(PACKAGE_RPC, "serverSi
 
 val byteArrayListEquals = MemberName(PACKAGE_UTIL, "byteArrayListsEqual")
 
-private val JSPB = ClassName(PACKAGE_BASE, "JSPB")
-val JSPB_BINARY_WRITER = JSPB.nestedClass("BinaryWriter")
+private val JSPB = ClassName(PACKAGE_EXTERNAL, "JSPB")
 val JSPB_BINARY_READER = JSPB.nestedClass("BinaryReader")
