+++ kmp-grpc-plugin/src/main/java/io/github/timortel/kmpgrpc/plugin/sourcegeneration/generators/service/ActualProtoServiceWriter.kt
@@ -10,7 +10,7 @@ abstract class ActualProtoServiceWriter : ProtoServiceWriter(true) {
     abstract val callOptionsType: TypeName
 
     /**
-    * Create empty call options for your platform, e.g. CallOptions.empty() or CallOptions()
+    * Create empty call options for your platform, e.g., CallOptions.empty() or CallOptions()
     */
     abstract val createEmptyCallOptionsCode: CodeBlock
